feature: create handleKeySig interaction function
