+++ app/lib/buildKeySignature.ts
@@ -16,6 +16,5 @@ export const buildKeySignature = (
         .setContext(context)
         .setStave(stave)
         .render(context, glyphInfo.xPosition, glyphInfo.yPosition);
-      console.log(glyph);
     });
 };
