+++ app/lib/createBlankStaves.ts
@@ -17,11 +17,12 @@ const createBlankStaves = (
     keySig,
   } = params
 ): BlankStaves => {
+  const spaceAboveStaff = { space_above_staff_ln: -10 };
   const stavesArray: InstanceType<typeof Stave>[] = [];
   if (regularStaveWidth)
     for (let i = 0; i < numStaves; i++) {
       let staveWidth = i === 0 ? firstStaveWidth : regularStaveWidth;
-      let stave = new Stave(x, y, staveWidth);
+      let stave = new Stave(x, y, staveWidth, spaceAboveStaff);
       i === 0 && clef ? stave.addClef(clef) : null;
       i === 0 && timeSig ? stave.addTimeSignature(timeSig) : null;
       i === 0 && keySig ? stave.addKeySignature(keySig) : null;
