+++ app/lib/keySigArray.ts
@@ -1,0 +1,1 @@
+export const keySigArray = ["g/5", "f/5", "e/5", "d/5", "c/5", "b/4", "a/4"];
