--- app/lib/setUpRendererAndDrawNotesOld.ts
@@ -1,45 +1,0 @@
-import VexFlow from "vexflow";
-import createBlankStaves from "../lib/createBlankStaves";
-import { RenderStavesAndNotesParamsOld } from "./typesAndInterfaces";
-const { Formatter } = VexFlow.Flow;
-
-export const setupRendererAndDrawNotesOld = (
-  params: RenderStavesAndNotesParamsOld
-): void => {
-  const {
-    rendererRef,
-    font,
-    fontSize,
-    numStaves,
-    rendererWidth,
-    rendererHeight,
-    yPositionOfStaves,
-    xPositionOfStaves,
-    clef,
-    firstStaveWidth,
-    keySig,
-    setStaves,
-    notesData,
-    staves: blankStaves,
-  } = params;
-  const renderer = rendererRef?.current;
-  renderer?.resize(rendererWidth, rendererHeight);
-  const context = renderer && renderer.getContext();
-  context?.setFont(font, fontSize);
-  context?.clear();
-  if (context && rendererRef) {
-    setStaves(() =>
-      createBlankStaves({
-        numStaves,
-        context,
-        firstStaveWidth,
-        x: xPositionOfStaves,
-        y: yPositionOfStaves,
-        regularStaveWidth: 300,
-        clef,
-        keySig,
-      })
-    );
-  }
- 
-};
