+++ app/lib/setupRendererAndDrawNotes.ts
