--- app/lib/setupRendererAndDrawNotesNew.ts
@@ -1,57 +1,0 @@
-import VexFlow from "vexflow";
-import createBlankStavesNew from "./createBlankStavesNew";
-import { RenderStavesAndNotesParamsOld } from "./typesAndInterfaces";
-const { Formatter } = VexFlow.Flow;
-
-export const setupRendererAndDrawNotesNew = (
-  params: RenderStavesAndNotesParamsOld
-): void => {
-  const {
-    rendererRef,
-    font,
-    fontSize,
-    numStaves,
-    rendererWidth,
-    rendererHeight,
-    yPositionOfStaves,
-    xPositionOfStaves,
-    clef,
-    firstStaveWidth,
-    keySig,
-    setStaves,
-    notesData,
-    staves: blankStaves,
-  } = params;
-
-  const renderer = rendererRef?.current;
-  renderer?.resize(rendererWidth, rendererHeight);
-  const context = renderer && renderer.getContext();
-  context?.setFont(font, fontSize);
-  context?.clear();
-
-  if (context && rendererRef) {
-    setStaves(() =>
-      createBlankStavesNew({
-        numStaves,
-        context,
-        firstStaveWidth,
-        x: xPositionOfStaves,
-        y: yPositionOfStaves,
-        regularStaveWidth: 300,
-        clef,
-        keySig,
-      })
-    );
-  }
-
-  notesData &&
-    notesData.forEach((barData, index) => {
-      if (barData) {
-        const staveNotes = barData.map(({ newStaveNote }) => newStaveNote);
-        if (staveNotes.length > 0) {
-          context &&
-            Formatter.FormatAndDraw(context, blankStaves[index], staveNotes);
-        }
-      }
-    });
-};
