+++ app/lib/typesAndInterfaces.ts
@@ -122,7 +122,7 @@ export interface NoteStringData {
 }
 export interface NotesAndCoordinatesData {
   note: string;
-  originalNote: string,
+  originalNote: string;
   yCoordinateMin: number;
   yCoordinateMax: number;
   userClickY?: number;
@@ -165,27 +165,10 @@ export interface RenderStavesAndNotesParams {
   regularStaveWidth?: number | null;
   setStaves: SetStaves;
   scaleDataMatrix?: ScaleData[][];
-  staves: BlankStaves;
-}
-
-export interface RenderStavesAndNotesParamsOld {
-  rendererRef: RendererRef | null;
-  font: string;
-  fontSize: number;
-  numStaves: number;
-  rendererWidth: number;
-  rendererHeight: number;
-  yPositionOfStaves: number;
-  xPositionOfStaves: number;
-  clef: string;
-  timeSig?: string;
-  keySig?: string;
-  firstStaveWidth: number;
-  regularStaveWidth?: number | null;
-  setStaves: SetStaves;
   notesData?: NoteData | null;
   staves: BlankStaves;
 }
+
 export interface RenderStaves {
   rendererRef: RendererRef | null;
   font: string;
